LayoutTests/inspector/dom-debugger/attribute-modified-style-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests functionality of attribute modified DOM breakpoints when modifying the inline style attribute.


== Running test suite: DOMBreakpoint.AttributeModified
-- Running test case: DOMBreakpoint.AttributeModified.Style.Add
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...
Getting 'display'...
PASS: 'display' should have the value ''.
Setting 'display: none;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value ''.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.
-- Running test teardown.

-- Running test case: DOMBreakpoint.AttributeModified.Style.ReplaceSame
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...
Getting 'display'...
PASS: 'display' should have the value 'none'.
Setting 'display: none;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'none'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.
-- Running test teardown.

-- Running test case: DOMBreakpoint.AttributeModified.Style.ReplaceDifferent
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...
Getting 'display'...
PASS: 'display' should have the value 'none'.
Setting 'display: block;'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'none'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'block'.
-- Running test teardown.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Remove
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...
Getting 'display'...
PASS: 'display' should have the value 'block'.
Removing 'display'...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value 'block'.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value ''.
-- Running test teardown.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Text
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...
Getting 'display'...
PASS: 'display' should have the value ''.
Setting style text to ''...
PASS: Paused.
Getting 'display'...
PASS: 'display' should have the value ''.
PASS: Resumed.
Getting 'display'...
PASS: 'display' should have the value 'none'.
-- Running test teardown.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Options.Condition
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Options.IgnoreCount
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Options.Action.Log
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMBreakpoint.AttributeModified.Style.Options.Actions.Evaluate
Adding "attribute-modified:1,HTML,1,BODY,1,DIV" DOM Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
